ui(testutils): add CommandServiceTrigger
