+++ packages/app/package.json
@@ -171,12 +171,11 @@
     "@testing-library/user-event": "^13.1.9",
     "@types/connect-redis": "^0.0.15",
     "@types/es-module-lexer": "^0.3.0",
+    "@types/faker": "5.5.7",
     "@types/file-saver": "^2.0.2",
     "@types/jest": "^26.0.8",
     "@types/minio": "^7.0.6",
     "@types/redis": "^2.8.28",
-    "faker": "5.5.3",
-    "@types/faker": "5.5.3",
     "@types/supertest": "^2.0.11",
     "@types/testing-library__jest-dom": "^5.9.5",
     "@types/testing-library__react": "^10.2.0",
@@ -211,6 +210,7 @@
     "eslint-plugin-react-hooks": "^4.2.0",
     "eslint-plugin-testing-library": "^4.6.0",
     "eslint-webpack-plugin": "^2.5.2",
+    "faker": "5.5.3",
     "file-loader": "^6.0.0",
     "fs-extra": "^8.1.0",
     "graphql-codegen-golang": "^0.1.2",
